Added cashu token receiving
