+++ components/utility-components/redeem-button.tsx
@@ -1,4 +1,4 @@
-aimport { useState, useEffect, useContext, useMemo } from "react";
+import { useState, useEffect, useContext, useMemo } from "react";
 import Link from "next/link";
 import { Button, Spinner, Tooltip } from "@nextui-org/react";
 import { useTheme } from "next-themes";
