+++ pages/wallet/index.tsx
@@ -26,134 +26,34 @@ import {
 import { SHOPSTRBUTTONCLASSNAMES } from "../../components/utility/STATIC-VARIABLES";
 import { getLocalStorageData } from "../../components/utility/nostr-helper-functions";
 import { captureInvoicePaidmetric } from "../../components/utility/metrics-helper-functions";
+import MintButton from "../../components/wallet/mint-button";
+import ReceiveButton from "../../components/wallet/receive-button"
 import { CashuMint, CashuWallet, getEncodedToken } from "@cashu/cashu-ts";
 
 const Wallet = () => {
-  const [showMintModal, setShowMintModal] = useState(false);
-  const [showInvoiceCard, setShowInvoiceCard] = useState(false);
+  const [totalBalance, setTotalBalance] = useState(0);
+  
+  const { tokens } = getLocalStorageData();
 
-  const [paymentConfirmed, setPaymentConfirmed] = useState(false);
-  const [qrCodeUrl, setQrCodeUrl] = useState<string | null>(null);
-  const [invoice, setInvoice] = useState("");
-  const [copiedToClipboard, setCopiedToClipboard] = useState(false);
-
-  const { mints, tokens } = getLocalStorageData();
-
-  const {
-    handleSubmit: handleMintSubmit,
-    formState: { errors },
-    control: mintControl,
-    reset: mintReset,
-  } = useForm();
-
-  const handleToggleMintModal = () => {
-    mintReset();
-    setShowMintModal(!showMintModal);
-    setShowInvoiceCard(false);
-  };
-
-  const onMintSubmit = async (data: { [x: string]: any }) => {
-    let numSats = data["sats"];
-    setShowInvoiceCard(true);
-    await handleMint(numSats);
-  };
-
-  const handleMint = async (numSats: number) => {
-    const wallet = new CashuWallet(new CashuMint(mints[0]));
-
-    const mintInvoice = await axios.post("/api/cashu/request-mint", {
-      mintUrl: mints[0],
-      total: numSats,
-      currency: "SATS",
-    });
-
-    const { id, pr, hash } = mintInvoice.data;
-
-    setInvoice(pr);
-
-    const QRCode = require("qrcode");
-
-    QRCode.toDataURL(pr)
-      .then((url: string) => {
-        setQrCodeUrl(url);
-      })
-      .catch((err: any) => {
-        console.error("ERROR", err);
-      });
-
-    invoiceHasBeenPaid(wallet, numSats, hash, id);
-  };
-
-  /** CHECKS WHETHER INVOICE HAS BEEN PAID */
-  async function invoiceHasBeenPaid(
-    wallet: CashuWallet,
-    numSats: number,
-    hash: string,
-    metricsInvoiceId: string,
-  ) {
-    let encoded;
-
-    while (true) {
-      try {
-        const { proofs } = await wallet.requestTokens(numSats, hash);
-
-        // Encoded proofs can be spent at the mint
-        encoded = getEncodedToken({
-          token: [
-            {
-              mint: mints[0],
-              proofs,
-            },
-          ],
-        });
-
-        if (encoded) {
-          localStorage.setItem(
-            "tokens",
-            JSON.stringify([tokens[0], { mint: mints[0], proofs }]),
-          );
-          captureInvoicePaidmetric(metricsInvoiceId, productData.id);
-          setPaymentConfirmed(true);
-          setQrCodeUrl(null);
-          setTimeout(() => {
-            router.push("/"); // takes you back to the home page after payment has been confirmed by cashu mint api
-          }, 1900); // 1.9 seconds is the amount of time for the checkmark animation to play
-          handleToggleMintModal();
-          break;
-        }
-      } catch (error) {
-        console.error(error);
-
-        await new Promise((resolve) => setTimeout(resolve, 2000));
-      }
-    }
-  }
-
-  const handleCopyInvoice = () => {
-    navigator.clipboard.writeText(invoice);
-    setCopiedToClipboard(true);
-    // after 2 seconds, set copiedToClipboard back to false
-    setTimeout(() => {
-      setCopiedToClipboard(false);
-    }, 2000);
-  };
+  useEffect(() => {
+    const tokensTotal = tokens && tokens.length > 0
+      ? tokens.reduce((acc, current) => {
+          const proofsTotal = current.proofs ? current.proofs.reduce((acc, proof) => acc + proof.amount, 0) : 0;
+          return acc + proofsTotal;
+        }, 0)
+      : 0;
+    setTotalBalance(tokensTotal);
+  }, [tokens]);
 
   return (
     <div className="flex min-h-screen flex-col bg-light-bg pb-20 pt-4 dark:bg-dark-bg sm:ml-[120px] md:ml-[250px]">
       <center>
         <p className="break-words text-center text-6xl text-xl text-light-text dark:text-dark-text">
-          0 sats
+          {totalBalance} sats
         </p>
       </center>
       <center>
-        <Button
-          className={SHOPSTRBUTTONCLASSNAMES + " m-2"}
-          startContent={
-            <ArrowDownTrayIcon className="h-6 w-6 hover:text-yellow-500" />
-          }
-        >
-          Receive
-        </Button>
+        <ReceiveButton />
         <Button
           className={SHOPSTRBUTTONCLASSNAMES + " m-2"}
           startContent={
@@ -164,15 +64,7 @@ const Wallet = () => {
         </Button>
       </center>
       <center>
-        <Button
-          className={SHOPSTRBUTTONCLASSNAMES + " m-2"}
-          onClick={() => setShowMintModal(!showMintModal)}
-          startContent={
-            <BanknotesIcon className="h-6 w-6 hover:text-yellow-500" />
-          }
-        >
-          Mint
-        </Button>
+        <MintButton />
         <Button
           className={SHOPSTRBUTTONCLASSNAMES + " m-2"}
           startContent={<BoltIcon className="h-6 w-6 hover:text-yellow-500" />}
@@ -180,146 +72,6 @@ const Wallet = () => {
           Redeem
         </Button>
       </center>
-      <Modal
-        backdrop="blur"
-        isOpen={showMintModal}
-        onClose={handleToggleMintModal}
-        classNames={{
-          body: "py-6",
-          backdrop: "bg-[#292f46]/50 backdrop-opacity-60",
-          // base: "border-[#292f46] bg-[#19172c] dark:bg-[#19172c] text-[#a8b0d3]",
-          header: "border-b-[1px] border-[#292f46]",
-          footer: "border-t-[1px] border-[#292f46]",
-          closeButton: "hover:bg-black/5 active:bg-white/10",
-        }}
-        scrollBehavior={"outside"}
-        size="2xl"
-      >
-        <ModalContent>
-          <ModalHeader className="flex flex-col gap-1 text-light-text dark:text-dark-text">
-            Mint Tokens
-          </ModalHeader>
-          <form onSubmit={handleMintSubmit(onMintSubmit)}>
-            <ModalBody>
-              <Controller
-                name="sats"
-                control={mintControl}
-                rules={{
-                  required: "A whole number is required.",
-                  maxLength: {
-                    value: 300,
-                    message: "This input exceed maxLength of 300.",
-                  },
-                  validate: (value) =>
-                    /^\d+$/.test(value) || "Please enter a whole number.",
-                }}
-                render={({
-                  field: { onChange, onBlur, value },
-                  fieldState: { error },
-                }) => {
-                  let isErrored = error !== undefined;
-                  let errorMessage: string = error?.message
-                    ? error.message
-                    : "";
-                  return (
-                    <Input
-                      className="text-light-text dark:text-dark-text"
-                      autoFocus
-                      variant="bordered"
-                      fullWidth={true}
-                      label="Token amount in sats"
-                      labelPlacement="inside"
-                      isInvalid={isErrored}
-                      errorMessage={errorMessage}
-                      // controller props
-                      onChange={onChange} // send value to hook form
-                      onBlur={onBlur} // notify when input is touched/blur
-                      value={value}
-                    />
-                  );
-                }}
-              />
-              {showInvoiceCard && (
-                <Card className="mt-3 max-w-[700px]">
-                  <CardHeader className="flex justify-center gap-3">
-                    <span className="text-xl font-bold">Lightning Invoice</span>
-                  </CardHeader>
-                  <Divider />
-                  <CardBody className="flex flex-col items-center">
-                    {!paymentConfirmed ? (
-                      <div className="flex flex-col items-center justify-center">
-                        {qrCodeUrl ? (
-                          <>
-                            <Image
-                              alt="Lightning invoice"
-                              className="object-cover"
-                              src={qrCodeUrl}
-                            />
-                            <div className="flex items-center justify-center">
-                              <p className="text-center">
-                                {invoice.length > 30
-                                  ? `${invoice.substring(
-                                      0,
-                                      10,
-                                    )}...${invoice.substring(
-                                      invoice.length - 10,
-                                      invoice.length,
-                                    )}`
-                                  : invoice}
-                              </p>
-                              <ClipboardIcon
-                                onClick={handleCopyInvoice}
-                                className={`ml-2 h-4 w-4 cursor-pointer ${
-                                  copiedToClipboard ? "hidden" : ""
-                                }`}
-                              />
-                              <CheckIcon
-                                className={`ml-2 h-4 w-4 cursor-pointer ${
-                                  copiedToClipboard ? "" : "hidden"
-                                }`}
-                              />
-                            </div>
-                          </>
-                        ) : (
-                          <div>
-                            <p>Waiting for lightning invoice...</p>
-                          </div>
-                        )}
-                      </div>
-                    ) : (
-                      <div className="flex flex-col items-center justify-center">
-                        <h3 className="mt-3 text-center text-lg font-medium leading-6 text-gray-900">
-                          Payment confirmed!
-                        </h3>
-                        <Image
-                          alt="Payment Confirmed"
-                          className="object-cover"
-                          src="../../payment-confirmed.gif"
-                          width={350}
-                        />
-                      </div>
-                    )}
-                  </CardBody>
-                </Card>
-              )}
-            </ModalBody>
-
-            <ModalFooter>
-              <Button
-                color="danger"
-                variant="light"
-                onClick={handleToggleMintModal}
-              >
-                Cancel
-              </Button>
-
-              <Button className={SHOPSTRBUTTONCLASSNAMES} type="submit">
-                Mint
-              </Button>
-            </ModalFooter>
-          </form>
-        </ModalContent>
-      </Modal>
     </div>
   );
 };
